requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
seaborn==0.13.2
matplotlib==3.8.2
scipy==1.12.0
numpy==1.26.3
reportlab==4.1.0
python-dotenv==1.0.1
